+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fixed an issue on Android where the build process would pick up version from the snapshot repository instead of the release repository, causing build failures.
+- Fixed an issue on iOS & Android where, when using `<StrictMode>`, the player's list of text tracks would be empty.
 
 ## [10.11.0] - 26-03-10